CP remove tab
TASK [CAPTIVE PORTAL] **********************************************************
ERROR! Syntax Error while loading YAML.
  found character '\t' that cannot start any token

The error appears to be in '/opt/iiab/iiab/roles/captiveportal/tasks/main.yml': line 27, column 41, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

- name: Enable or Disable Captive Portal
                                        ^ here
There appears to be a tab character at the start of the line.

YAML does not use tabs for formatting. Tabs should be replaced with spaces.